[AM-269] refactor: 이벤트용 Entity 디렉토리 구조 변경 (#31)
